+++ src/main/java/org/clarksnut/datasource/peru/PeruDebitNoteDatasource.java
@@ -26,9 +26,9 @@ public class PeruDebitNoteDatasource implements Datasource {
 
     private InformacionAdicionalBean informacionAdicional;
 
-    private Float totalVenta;
-    private Float totalDescuentoGlobal;
-    private Float totalOtrosCargos;
+    private Double totalVenta;
+    private Double totalDescuentoGlobal;
+    private Double totalOtrosCargos;
 
     private TributosBean tributos;
 
@@ -130,27 +130,27 @@ public class PeruDebitNoteDatasource implements Datasource {
         this.informacionAdicional = informacionAdicional;
     }
 
-    public Float getTotalVenta() {
+    public Double getTotalVenta() {
         return totalVenta;
     }
 
-    public void setTotalVenta(Float totalVenta) {
+    public void setTotalVenta(Double totalVenta) {
         this.totalVenta = totalVenta;
     }
 
-    public Float getTotalDescuentoGlobal() {
+    public Double getTotalDescuentoGlobal() {
         return totalDescuentoGlobal;
     }
 
-    public void setTotalDescuentoGlobal(Float totalDescuentoGlobal) {
+    public void setTotalDescuentoGlobal(Double totalDescuentoGlobal) {
         this.totalDescuentoGlobal = totalDescuentoGlobal;
     }
 
-    public Float getTotalOtrosCargos() {
+    public Double getTotalOtrosCargos() {
         return totalOtrosCargos;
     }
 
-    public void setTotalOtrosCargos(Float totalOtrosCargos) {
+    public void setTotalOtrosCargos(Double totalOtrosCargos) {
         this.totalOtrosCargos = totalOtrosCargos;
     }
